LPS-205650 Extract ScreenNavigationEntry methods from CountryDetailsScreenNavigationCategory to CountryDetailsScreenNavigationEntry, extends CountryDetailsScreenNavigationCategory in it and return getCategoryKey in getEntryKey method
