+++ modules/apps/address/address-web/src/main/java/com/liferay/address/web/internal/frontend/taglib/servlet/taglib/CountryDetailsScreenNavigationCategory.java
@@ -7,18 +7,9 @@ package com.liferay.address.web.internal.frontend.taglib.servlet.taglib;
 
 import com.liferay.address.web.internal.constants.CountryScreenNavigationConstants;
 import com.liferay.frontend.taglib.servlet.taglib.ScreenNavigationCategory;
-import com.liferay.frontend.taglib.servlet.taglib.ScreenNavigationEntry;
-import com.liferay.frontend.taglib.servlet.taglib.util.JSPRenderer;
 import com.liferay.portal.kernel.language.Language;
-import com.liferay.portal.kernel.model.Country;
-
-import java.io.IOException;
 
 import java.util.Locale;
-
-import javax.servlet.ServletContext;
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
 
 import org.osgi.service.component.annotations.Component;
 import org.osgi.service.component.annotations.Reference;
@@ -27,14 +18,11 @@ import org.osgi.service.component.annotations.Reference;
  * @author Pei-Jung Lan
  */
 @Component(
-	property = {
-		"screen.navigation.category.order:Integer=10",
-		"screen.navigation.entry.order:Integer=10"
-	},
-	service = {ScreenNavigationCategory.class, ScreenNavigationEntry.class}
+	property = "screen.navigation.category.order:Integer=10",
+	service = ScreenNavigationCategory.class
 )
 public class CountryDetailsScreenNavigationCategory
-	implements ScreenNavigationCategory, ScreenNavigationEntry<Country> {
+	implements ScreenNavigationCategory {
 
 	@Override
 	public String getCategoryKey() {
@@ -42,13 +30,8 @@ public class CountryDetailsScreenNavigationCategory
 	}
 
 	@Override
-	public String getEntryKey() {
-		return CountryScreenNavigationConstants.CATEGORY_KEY_DETAILS;
-	}
-
-	@Override
 	public String getLabel(Locale locale) {
-		return _language.get(locale, "details");
+		return language.get(locale, "details");
 	}
 
 	@Override
@@ -56,24 +39,7 @@ public class CountryDetailsScreenNavigationCategory
 		return CountryScreenNavigationConstants.SCREEN_NAVIGATION_KEY_COUNTRY;
 	}
 
-	@Override
-	public void render(
-			HttpServletRequest httpServletRequest,
-			HttpServletResponse httpServletResponse)
-		throws IOException {
-
-		_jspRenderer.renderJSP(
-			_servletContext, httpServletRequest, httpServletResponse,
-			"/country/details.jsp");
-	}
-
 	@Reference
-	private JSPRenderer _jspRenderer;
-
-	@Reference
-	private Language _language;
-
-	@Reference(target = "(osgi.web.symbolicname=com.liferay.address.web)")
-	private ServletContext _servletContext;
+	protected Language language;
 
 }
